pyproject.toml
[tool.poetry]
name = "truepositive-test-task"
version = "0.1.0"
description = ""
authors = ["valiaisnotprogrammer <ValiaChernykh@yandex.ru>"]
#readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.4"
djangorestframework = "^3.15.1"
selenium = "^4.20.0"
aiogram = "^3.5.0"
psycopg2-binary = "^2.9.9"
requests = "^2.31.0"
loguru = "^0.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
